cut & del icons improvements

+++ client/offer/items/DeleteIcon.tsx
@@ -1,3 +1,4 @@
+import { saveOfferIntoLocalStorage } from 'client/modules/localStorage'
 import { useDispatchTyped, useSelectorTyped } from 'client/store'
 import { ItemType } from 'client/types'
 import { RxCross2 } from 'react-icons/rx'
@@ -9,14 +10,14 @@ type Props = {
 
 export const DeleteIcon = ({ itemToDelete }: Props) => {
   const dispatch = useDispatchTyped()
-  const isPasteMode = useSelectorTyped(state => state.copy.isShown)
+  const isLastItem = useSelectorTyped(state => state.offer.items.filter((item) => item.type !== 'paste').length === 1)
 
   return (
     <RxCross2
       css={{
-        color: isPasteMode ? '#acacac' : 'inherit',
-        cursor: isPasteMode ? 'default' : 'pointer',
-        ...(!isPasteMode && {
+        color: isLastItem ? '#acacac' : 'inherit',
+        cursor: isLastItem ? 'default' : 'pointer',
+        ...(!isLastItem && {
           ':hover': {
             scale: '1.3',
             color: 'black',
@@ -25,7 +26,9 @@ export const DeleteIcon = ({ itemToDelete }: Props) => {
         })
       }}
       onClick={() => {
+        if (isLastItem) return
         dispatch(deleteItem(itemToDelete))
+        saveOfferIntoLocalStorage()
       }}
     />
   )
